refactor: AddPostDialog 분리
